Extract var camel casing to function which will perform all var name normalization
