+++ packages/destination-actions/src/destinations/fullstory/identifyUser/index.ts
@@ -1,7 +1,7 @@
 import type { ActionDefinition } from '@segment/actions-core'
-import camelCase from 'lodash/camelCase'
 import type { Settings } from '../generated-types'
 import { setUserPropertiesRequestParams } from '../request-params'
+import { normalizeVarNames } from '../vars'
 
 const action: ActionDefinition<Settings> = {
   title: 'Identify User',
@@ -57,25 +57,16 @@ const action: ActionDefinition<Settings> = {
   },
   perform: (request, { payload, settings }) => {
     const { traits, anonymousId, userId, email, displayName } = payload
-    let newTraits: Record<string, unknown> = {}
 
-    if (traits) {
-      newTraits = Object.entries(traits).reduce(
-        (acc, [key, value]) => ({
-          ...acc,
-          [camelCaseField(key)]: value
-        }),
-        {}
-      )
-    }
+    const normalizedTraits = normalizeVarNames(traits, { camelCase: true })
 
     if (anonymousId) {
-      newTraits.segmentAnonymousId_str = anonymousId
+      normalizedTraits.segmentAnonymousId_str = anonymousId
     }
 
     // TODO(nate): Include a source value once the HTTP API is updated to support it
     const requestBody = {
-      ...newTraits,
+      ...normalizedTraits,
       ...(email !== undefined && { email }),
       ...(displayName !== undefined && { displayName })
     }
@@ -83,39 +74,6 @@ const action: ActionDefinition<Settings> = {
     const { url, options } = setUserPropertiesRequestParams(settings, userId, requestBody)
     return request(url, options)
   }
-}
-
-/**
- * Camel cases `.`, `-`, `_`, and white space within fieldNames. Leaves type suffix alone.
- *
- * NOTE: Does not fix otherwise malformed fieldNames.
- * FullStory will scrub characters from keys that do not conform to /^[a-zA-Z][a-zA-Z0-9_]*$/.
- *
- * @param {string} fieldName
- */
-function camelCaseField(fieldName: string) {
-  // Do not camel case across type suffixes.
-  const parts = fieldName.split('_')
-  if (parts.length > 1) {
-    const typeSuffix = parts.pop()
-    switch (typeSuffix) {
-      case 'str':
-      case 'int':
-      case 'date':
-      case 'real':
-      case 'bool':
-      case 'strs':
-      case 'ints':
-      case 'dates':
-      case 'reals':
-      case 'bools':
-        return camelCase(parts.join('_')) + '_' + typeSuffix
-      default: // passthrough
-    }
-  }
-
-  // No type suffix found. Camel case the whole field name.
-  return camelCase(fieldName)
 }
 
 export default action
